Refactored course price API frontend code to namespace on username
This is similar to the changes I made in #2692 to the dashboard reducer
code. Basically, we want to namespace the returned values from the
course price API on username, so that we can store values for multiple
users at the same time (necessary to support the changes we need to make
for the staff learner info view).

closes #2923
pr #2924
