+++ static/js/actions/index.js
@@ -7,7 +7,6 @@ import * as api from '../lib/api';
 import type { CheckoutResponse } from '../flow/checkoutTypes';
 import type { APIErrorInfo } from '../flow/generalTypes';
 import type { Dispatcher } from '../flow/reduxTypes';
-import type { CoursePrices } from '../flow/dashboardTypes';
 
 // constants for fetch status (these are not action types)
 export const FETCH_FAILURE = 'FETCH_FAILURE';
@@ -45,28 +44,3 @@ export const RECEIVE_CHECKOUT_FAILURE = 'RECEIVE_CHECKOUT_FAILURE';
 export const receiveCheckoutFailure = createAction(RECEIVE_CHECKOUT_FAILURE,
   (errorInfo: APIErrorInfo) => ({errorInfo })
 );
-
-// course price actions
-export const REQUEST_COURSE_PRICES = 'REQUEST_COURSE_PRICES';
-export const requestCoursePrices = createAction(REQUEST_COURSE_PRICES);
-
-export const RECEIVE_COURSE_PRICES_SUCCESS = 'RECEIVE_COURSE_PRICES_SUCCESS';
-export const receiveCoursePricesSuccess = createAction(RECEIVE_COURSE_PRICES_SUCCESS);
-
-export const RECEIVE_COURSE_PRICES_FAILURE = 'RECEIVE_COURSE_PRICES_FAILURE';
-export const receiveCoursePricesFailure = createAction(RECEIVE_COURSE_PRICES_FAILURE);
-
-export const CLEAR_COURSE_PRICES = 'CLEAR_COURSE_PRICES';
-export const clearCoursePrices = createAction(CLEAR_COURSE_PRICES);
-
-export function fetchCoursePrices(): Dispatcher<CoursePrices> {
-  return (dispatch: Dispatch) => {
-    dispatch(requestCoursePrices());
-    return api.getCoursePrices().
-      then(coursePrices => dispatch(receiveCoursePricesSuccess(coursePrices))).
-      catch(error => {
-        dispatch(receiveCoursePricesFailure(error));
-        // the exception is assumed handled and will not be propagated
-      });
-  };
-}
