+++ static/js/actions/index_test.js
@@ -5,32 +5,14 @@ import {
   requestCheckout,
   receiveCheckoutSuccess,
   receiveCheckoutFailure,
-  requestCoursePrices,
-  receiveCoursePricesSuccess,
-  receiveCoursePricesFailure,
-  clearCoursePrices,
 
   REQUEST_CHECKOUT,
   RECEIVE_CHECKOUT_SUCCESS,
   RECEIVE_CHECKOUT_FAILURE,
-  REQUEST_COURSE_PRICES,
-  RECEIVE_COURSE_PRICES_SUCCESS,
-  RECEIVE_COURSE_PRICES_FAILURE,
-  CLEAR_COURSE_PRICES,
 } from './';
-import { assertCreatedActionHelper } from './test_util';
 import { ERROR_RESPONSE } from '../constants';
 
 describe('generated index action helpers', () => {
-  it('should create all action creators', () => {
-    [
-      [requestCoursePrices, REQUEST_COURSE_PRICES],
-      [receiveCoursePricesSuccess, RECEIVE_COURSE_PRICES_SUCCESS],
-      [receiveCoursePricesFailure, RECEIVE_COURSE_PRICES_FAILURE],
-      [clearCoursePrices, CLEAR_COURSE_PRICES],
-    ].forEach(assertCreatedActionHelper);
-  });
-
   it('requestCheckout passes a course id', () => {
     assert.deepEqual(requestCheckout('course_id'), {
       type: REQUEST_CHECKOUT,
