+++ static/js/reducers/course_enrollments_test.js
@@ -14,7 +14,7 @@ import {
   RECEIVE_ADD_COURSE_ENROLLMENT_FAILURE,
 } from '../actions/course_enrollments';
 import * as api from '../lib/api';
-import * as actions from '../actions';
+import * as coursePriceActions from '../actions/course_prices';
 import * as dashboardActions from '../actions/dashboard';
 import rootReducer from '../reducers';
 
@@ -36,7 +36,7 @@ describe('enrollments', () => {
     beforeEach(() => {
       dispatchThen = store.createDispatchThen(state => state.courseEnrollments);
 
-      fetchCoursePricesStub = sandbox.stub(actions, 'fetchCoursePrices');
+      fetchCoursePricesStub = sandbox.stub(coursePriceActions, 'fetchCoursePrices');
       fetchCoursePricesStub.returns({type: "fake"});
       fetchDashboardStub = sandbox.stub(dashboardActions, 'fetchDashboard');
       fetchDashboardStub.returns({type: "fake"});
