+++ static/js/reducers/programs_test.js
@@ -29,7 +29,7 @@ import {
   SET_CURRENT_PROGRAM_ENROLLMENT,
 } from '../actions/programs';
 import * as api from '../lib/api';
-import * as actions from '../actions';
+import * as coursePriceActions from '../actions/course_prices';
 import * as dashboardActions from '../actions/dashboard';
 import rootReducer from '../reducers';
 
@@ -57,7 +57,7 @@ describe('enrollments', () => {
     beforeEach(() => {
       dispatchThen = store.createDispatchThen(state => state.programs);
 
-      fetchCoursePricesStub = sandbox.stub(actions, 'fetchCoursePrices');
+      fetchCoursePricesStub = sandbox.stub(coursePriceActions, 'fetchCoursePrices');
       fetchCoursePricesStub.returns({type: "fake"});
       fetchDashboardStub = sandbox.stub(dashboardActions, 'fetchDashboard');
       fetchDashboardStub.returns({type: "fake"});
